requirements.txt
dj-database-url==0.5.0
Django==2.1
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
Markdown==2.6.11
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.0
psycopg2==2.6.2
